fix クラスの整理
