+++ app/src/main/java/work/airz/primanager/ISaveTicket.kt
@@ -1,0 +1,7 @@
+package work.airz.primanager
+
+interface ISaveTicket {
+    fun getStoredData()
+    fun saveData()
+}
+
